refac: moves the Role class in the Domain file.
